requirements.txt
google-cloud-storage==2.16.0
google-auth==2.28.2